Update ad-hoc licence to just licence (#2109)
* Update ad-hoc licence to just licence

https://eaflood.atlassian.net/browse/WATER-5101

We previously implemented 'ad-hoc' returns as a journey for notices. This is going to change.

Ad-hoc will now be known as licence, this page will still be the start of sending a returns notice for an individual licence.

This change removes the use of ad-hoc for the licence pages.

It does not remove the ad-hoc journey, that will be in a later change.

* fix: test paths