Add new base licences presenter with format method
When we came to fix the summary page we spotted that the code for formatting the amounts was in both places, exactly the same. The difference was that one handles many licence version purposes, whereas summary only cares if there is just one.

So, we left that handling logic where it was, but moved the formatting logic to a new `base-licences.presenter.js`.

Whilst we were at it, we also noticed that this is the one place in the service where we _don't_ format the numbers into a human readable string (commas for thousands etc).

We double checked with UCD, and they confirmed this was an oversight and on the TODO list to sort. Well, your welcome!