Update view licence summary tests
We also needed to tweak this one to expect human readable numbers. Plus the correction to the measure.

But this was also testing all permutations, which is now sufficiently covered by what we have done for the base-licence.presenter.js. So, we could simplify what was there to focus on what the presenter itself was doing.