Fix missed wrong abstraction amount measure text (#2110)
https://eaflood.atlassian.net/browse/WATER-5068

In [Fix instantaneous abstraction amount measure text](https://github.com/DEFRA/water-abstraction-system/pull/2103), we addressed an issue where the wrong text was displayed. We were showing 'cubic metres' when it should have been litres.

Or so we thought!

We missed where a licence has just one purpose with an instantaneous abstraction amount. This means that on the licence summary tab, we don't show the text 'Multiple abstractions'; instead, we show the amounts, and we're using the wrong measure.

This fixes that missed scenario.

Additionally, we also noticed (finally!) that the numbers we were showing were the only examples of non-human-readable numbers we could think of (as in commas for thousands).

A quick check with our UCD brethren confirmed this had been an oversight for a very long time. So, we fix that as well.