Add tests to clean empty bill runs
We want to be able to log if a task fails, but allow the others to run. Within the 'process' service that would mean wrapping each call in its own try/catch.

We think it would be better to follow a pattern we have used elsewhere, where it is the responsibility of the service itself not to return an error on failure. To that end we'll handle any errors thrown in the clean empty bill runs service.

This failing test confirms it is not yet handling errors.