Update recipient cancel for abstraction journey
https://eaflood.atlassian.net/browse/WATER-5067

There is existing logic to allow te user to cancel the notice on the recipients page. There is also a cancel option earlier in the abstraction alerts journey on the select licence matches page.

We need to allow the user to cancel the notice on for an abstraction alert on the recipients cancel page.

This change updates the existing cancel recipients  logic to handle the abstraction alerts journey.

It would have been more complicated to redirect to the existing alerts cancel page and set the correct back link navigation etc. So we have updated the existing logic to accommodate this.

Doing the change this way allows us to adapt to the potential design change in the future.

This cancel page uses the notice reference code the same as the other to signify we are pass the journey specific pages.