Update clean empty bill runs
Oh my oh my! We _really_ had it wrong on our first iteration of cleaning the empty bill runs!

When we recreated an empty presroc bill run, we realised that it will generate not just the bill run and the charge version years, but volumes, bills, bill licences, transactions, plus there is the bill run recorded in the Charging Module!

Fortunately, we've already done all the work to handle all this with our 'cancel bill run' feature.

So, we completely re-wrote this clean task to first identify the bill runs that are empty, then use this information to feed the existing delete bill run services, using the same process as the cancel bill run page.