Update clean empty void return logs
Like clean empty bill runs, we want to ensure we are always getting a count back. It is another way of ensuring that no matter what happens in the service, 'process' can expect the same result meaning there is less likely to be an issue that blocks the other tasks from running.