Handle empty pre-sroc bill runs in clean job  (#2112)
https://eaflood.atlassian.net/browse/WATER-5075

In [Add clean empty bill runs task to clean job](https://github.com/DEFRA/water-abstraction-system/pull/2082), we added the automatic cleaning of empty bill runs to our new 'clean' job.

All seemed well, so it was deployed in [release CHG0116581](https://github.com/DEFRA/water-abstraction-system/releases/tag/v0.35.0).

However, on its first run, it failed. We can see from the logs that it encountered an empty `PRESROC` bill run, which has child records, even when those child records are empty. But, this should have only caused the bill run step to fail. Instead, none of the other tasks ran.

So, first, we ensure that the tasks are isolated. Then we deal with empty `PRESROC` bill runs.

And oh my! We _really_ had it wrong on our first iteration of cleaning the empty bill runs!

When we recreated an empty `PRESROC` bill run, we realised that it would generate not just the bill run and the charge version years, but also volumes, bills, bill licences, transactions, and, additionally, the bill run is recorded in the Charging Module.

Fortunately, we've already done all the work to handle deleting all this with our 'cancel bill run' feature.

Therefore, we rewrote the entire clean empty bill runs task first to identify the empty bill runs, then used this information to feed the existing delete bill run services, following the same process as the cancel bill run page.