Update submit notice type to determine notice type
https://eaflood.atlassian.net/browse/WATER-5011

We previously implemented the 'select the notice page' for the setup ad-hoc journey. The ad-hoc journey is being removed and replaced with a new journey that is  functionality the same but with the use of the term 'ad-hoc' removed.

We will use the 'licenceRef' to determine if the journey is for a single licence or multiple. This will be updated in a later change.

This change sets the session data for the 'select the notice type' page. This works the same as when we set the initiate session data.