Refactor notice setup ad-hoc journey (#2121)
* Refactor notice setup ad-hoc journey

https://eaflood.atlassian.net/browse/WATER-5101

An 'ad-hoc' notice is currently considered its own journey in the existent code. This means a notice will be of type 'adHocReminder'.

This will no longer be the case. We will be introducing paper forms in later changes which also require selecting a specific licence ref to send a notice to.

This change removes the 'ad-hoc' journey. Were the journey key of 'ad-hoc' has been used, it has been replaced with checking if the 'licenceRef' exists.

This allows us to set the journey data for various notices going forwards.