Add notice setup returns for paper forms (#2130)
* Add notice setup returns for paper forms

https://eaflood.atlassian.net/browse/WATER-5023

As part of the update to the ad-hoc journey the user can select a notice type of for a paper for invitation.

This change adds the scaffolded code for returns for paper forms page.

We have not mentioned invitations, because of this changes in the future we want the code to flexible. We are concerned about the paper forms. If it is an invitation or changes to a reminder in the future it should have no relevance to the logic.