Fix notifications `view.njk` formatting (#2133)
It has been noticed during review of another PR https://github.com/DEFRA/water-abstraction-system/pull/2105 that in the view the `div` was in the wrong place. As that view was mostly a copy of `app/views/notifications/view.njk`. The issue also exists in this view.

This PR will fix the issue in `app/views/notifications/view.njk`.