Update notice setup check notice type to use change links
https://eaflood.atlassian.net/browse/WATER-5024

We scaffolded the 'check the notice type' page in a previous change and implemented a basic summary list.

This change adds the change links to the summary list, and implements the logic to handle a check page visited, consistent with the rest of the application.

We have also add flash notifications to be consistent with the rest of the service.