Support link from customer page to billing account (#2134)
https://eaflood.atlassian.net/browse/WATER-4988

We have recently migrated the legacy view billing account page from [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui) to here.

This has required updating links in the legacy UI to direct users to the new page. Only, as usual with the legacy service, finding all the links is problematic.

This change is because we have just spotted another one. This time from the **Customer** page (accessed from the licence summary contact details tab).

The **Customer** page is a standalone page. You can only access it from the link on the contact details tab, and it isn't licence-specific (a 'customer' can be linked to multiple licences).

Therefore, we feel obligated to allow the user to 'Go back...' to the **Customer** page from our new view billing account page. This means, along with updating the link in `water-abstraction-ui`, we also need to support getting the user back to it.