Amend `SubmitFullConditionService` to store licence condition display text
We want the check page to display the exact text of the licence condition, including the condition type, its number in the list (ie. the first radio button is condition 1, the second is condition 2, etc.), and the description and additional info. To get its number in the list, we need to fetch all of the conditions and match the selected condition then determine its index number. We could do this on the check page but it means fetching all conditions for the licence every time the page is rendered. To avoid this, we amend `SubmitFullConditionService` so that the conditions are fetched on submission and the required text stored in the session