Update notice setup notice type flash notification
https://eaflood.atlassian.net/browse/WATER-5024

We have implemented the 'check the notice type' page and added the change links.

This change adds a flash notification to the submit notice type service.

We have added a check to only show the notification when the notice type has changed.

This is handled by a previous change where the check page rendered the notification from the licence page.