Update notice setup return paper forms
https://eaflood.atlassian.net/browse/WATER-5023

We previously scaffolded the returns paper forms page.

This change updates the page to display the select options, the validation has been updated and the submit service.

When the user selects only one return we need to convert the payload option into an array to be used in the submit service. This has been implemented in the same established pattern.

Mock data has been used until we have a decision on the best way to extract the data.