set isUpload from return version flag (#2145)
https://eaflood.atlassian.net/browse/WATER-5117

A bug was found in production where external users were not able to see their due return logs if they were set up to use multiple upload. This was due to the code looking at the return requirement level at the is_upload column when it should be looking at the multiple_upload column on the return version.

This PR introduces a fix to use the correct place.