Handle return logs with no purposes
https://eaflood.atlassian.net/browse/WATER-5116

Our users got an error page when clicking into the returns tab for a licence. When we looked into the reason we found it was because some of the licence's return logs are missing keys details in their metadata property.

In the example reported it was the abstraction period, points, and purposes. But it was the lack of purposes breaking the tab.

How it happens is because the return log was generated from NALD data back when we imported it all returns information. Someone has managed to create a return version and requirement in NALD, then either omitted or deleted its purposes and points.

In NALD a return log would still be generated, and you could submit against it. But in WRLS this causes problems because we assume all records would have this information.

So, until we decide how to correct the return version and return logs, we need to be able to handle them and stop the UI from breaking.