Move formatting purpose logic to base presenter
Having fixed the returns tab (change is coming!) we then confirmed clicking one of these return logs with missing purposes caused the view return log page to error.

When we looked, we found the same logic for formatting a purpose there. So, rather than fix both places, we're moving the logic to the base-return-log presenter and we'll have both the returns tab and view return log page use the same helper.