Handle return logs with no purposes (#2146)
https://eaflood.atlassian.net/browse/WATER-5116

Our users got an error page when clicking into the returns tab for a licence. When we investigated the reason, we found that some of the licences' return logs were missing key details in their metadata property.

In the example reported, it was the abstraction period, points, and purposes that were missing. But it was the lack of purposes that broke the tab.

How it happens is because the return log was generated from NALD data back when we imported it, all returns information. Someone has managed to create a return version and requirement in NALD, then either omitted or deleted its purposes and points.

In NALD, a return log would still be generated, and you could submit against it. But in WRLS, this causes problems because we assume all records would have this information.

Therefore, until we determine how to correct the return version and return logs, we must be able to handle them and prevent the UI from breaking.