Add return id migration for return logs view (#2152)
https://eaflood.atlassian.net/browse/WATER-5023

We have a need for a unique id that is not based on the return.

A change has been made to returns to introduce this id here - https://github.com/DEFRA/water-abstraction-returns/pull/427

This change updates the public view for returns which id 'return_logs'. This id will be called 'return_id' with the intention of switching the columns / logic around in a later change.