Update notice setup returns paper forms change link
https://eaflood.atlassian.net/browse/WATER-5024

This change adds the returns data to the view and the associated change link.

Returns only shows for the 'paper-forms' journey and at least one will have been selected to show.

When multiple are selected we need to render the returns in a list in the UI. To achieve this we set the summary list value to html and add a <br> after each return.