Amend page/session names to match code
In several places we refer to `start-abstraction-period-day`, `start-abstraction-period-month`, `end-abstraction-period-day` and `end-abstraction-period-month` when reading data from the page and storing it in the session. However this is at adds with our convention in the code to refer to `abstractionPeriodStartDay`, `abstractionPeriodStartMonth`, etc. We therefore update our naming to change eg. `start-abstraction-period-day` to `abstraction-period-start-day` etc.