Create general abstraction period validation (#2154)
As part of the work on the [Enter Abstraction Period](https://github.com/DEFRA/water-abstraction-system/pull/2098) page, we have duplicated the abstraction period validation from the return versions journey -- usually this is acceptable as we wouldn't look to refactor until we have 3 such cases, but since we know we will have one more page coming shortly which will require abstraction period validation, we have decided to create a generic abstraction period validator now.

This ticket moves the abstraction period validator from `app/validators/return-versions/setup/` up to `app/validators/` and amends `SubmitAbstractionPeriodService` accordingly. While doing this we refactor the validator slightly to cut out some duplication, and we expand the unit test suite.

We also note a discrepancy between how we refer to things in templates and in code; in templates we refer to `start-abstraction-period-day`, `start-abstraction-period-month`, `end-abstraction-period-day` and `end-abstraction-period-month`, whereas in code we refer to `abstractionPeriodStartDay`, `abstractionPeriodStartMonth` etc. We therefore take the opportunity to rename the template naming to match, so we now have `abstraction-period-start-day` etc.