Remove the return version migration job
https://eaflood.atlassian.net/browse/WATER-4881

To support quarterly returns, we needed to set up new quarterly return versions for the water companies. We didn't want the business to have to do this manually, so we created a 'one-off' job to migrate the latest return version for each water company to a new one, that is just a copy, only it starts on 1 April 2025 and and has the 'quarterly' flag enabled.

That was run as part of our [release on 11 June 2025](https://eaflood.atlassian.net/projects/WATER/versions/23175) and was successful.

We'll never run that job again so we can now delete it from the app.