Update notice setup licence navigation (#2159)
https://eaflood.atlassian.net/browse/WATER-5024

This change updates the navigation for the ad-hoc journey for the licence page to align with our established patterns.

The back link will go back to the previous page, unless the user has completed the journey up to the check notice type page, then they will be redirected back that page.

When we navigate the user through the journey that may have selected the notice type already. When the licence changes we want the user to complete the journey again as this affects the generated returns that can be selected.