Update notice setup returns navigation
https://eaflood.atlassian.net/browse/WATER-5024

This change updates the navigation for the ad-hoc journey for the returns page to align with our established patterns.

The back link will go back to the previous page, unless the user has completed the journey up to the check notice type page, then they will be redirected back that page.