Update notice setup notice type navigation (#2161)
https://eaflood.atlassian.net/browse/WATER-5024

This change updates the navigation for the ad-hoc journey for the notice type page to align with our established patterns.

The back link will go back to the previous page, unless the user has completed the journey up to the check notice type page, then they will be redirected back that page.

When the user has navigated through the journey and changes to paper forms  from standard and continues to the new page, we need to clear the check visited flag to rest the journey.