Fix permissions for view notices page (#2165)
https://eaflood.atlassian.net/browse/WATER-5132

As part of switching returns management from NALD to WRLS, we needed to take control of sending return invitations and reminders.

Additionally, we needed to make changes to the View notices page (Notices page), which is why it was migrated.

Unfortunately, we didn't pay proper attention to which 'roles' had access to the legacy page, so users who used to be able to access the page can no longer.

This change updates the route's scope to match the legacy version of the page.