Update notice setup returns paper forms
https://eaflood.atlassian.net/browse/WATER-5023

This change updates the fetch returns due service for the returns paper forms page.

We only want to fetch return logs which are 'due' and before 'today'. We do not want to show future return logs.

This change also affects the submit licence service check where it checks there returns logs are to proceed with the journey.