Add select address service (#2144)
https://eaflood.atlassian.net/browse/WATER-5045

As part of the journey to create ad-hoc letters to be sent out to users we need to be able to select an address.

This PR adds in the functionality to query the OS places to return a list of addresses, choose one and have it saved to the session.