Amend how we parse out the result
We want to simplify access to the address facade results. So we add them as a property to the result object directly.

But we still want a copy of the body for those times when an error occurs. The message and the error will be properties of it and this allows us access to them.