Update Address lookup requests to convention (#2176)
https://eaflood.atlassian.net/browse/WATER-5119

We recently migrated the send return invitations and reminders notification journeys, which required sending requests to [GOV.UK Notify](https://www.notifications.service.gov.uk/) for the first time.

Building on that, we are adding the ability to send notices to specific licences and amending the address used. That meant talking to the [ea-address-facade](https://github.com/DEFRA/ea-address-facade).

We added the requests, but have only now just noted that they don't follow the conventions used when we first added 'requests' for the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) and the other legacy services to this project.

We have our `app/requests/base.request.js`, which handles using [Got](https://github.com/sindresorhus/got) to make HTTP requests and deal with the responses.

We then have 'base' service requests, for example, `app/requests/charging-module.request.js`. These handle setting the domain and any general operations, for example, applying authentication. They also control how exactly we can communicate with a service. Perhaps only `post` requests are supported, so only a Charging Module API `post` method is exposed.

Finally, you have the individual requests themselves. These are specific and are used to set the path to use and handle anything related to the request.

The **Notify** and **Address** requests have not followed this pattern. Instead, the specific requests are referencing `base.request.js` directly.

This change isn't critical. However, we have conventions and patterns, so it is best to follow them until we _purposely_ need to break them.

This change updates the address-based requests.