Update notice setup fetch download recipients
https://eaflood.atlassian.net/browse/WATER-5143

We previously implemented an update ad-hoc journey allowing the user to select a return to send the notice for.

This has highlighted an issue with our logic where we did not limit the returns logs due to 'today'. This means we were getting all the return logs due even future return logs.

This change update the 'FetchDownloadRecipientsService' to only return recipients where return logs are due and the endDate is in the pass.

To get to the page where this logic is due there is already a check on the return logs 'endDate'. This change aligns the two.