Remove existing /check endpoints
https://eaflood.atlassian.net/browse/WATER-5147

We've found an issue with the addresses our Notice engine is generating when sending letters via [GOV.UK Notify](https://www.notifications.service.gov.uk/).

There are lots of scenarios we want to check, which is where our `/check` endpoints come in. They allow us to quickly testing functionality in isolation knowing these endpoints aren't available in 'production'.

We want to add a `/check` endpoint for the changes were about to make to the notice address generation but spotted some redundant ones that are no longer used.

So, being a 'good team' we're deleting those first before adding our own!