Remove existing /check endpoints (#2183)
https://eaflood.atlassian.net/browse/WATER-5147

We've identified an issue with the addresses our Notice engine generates when sending letters via [GOV.UK Notify](https://www.notifications.service.gov.uk/).

There are lots of scenarios we want to check, which is where our `/check` endpoints come in. They allow us to quickly test functionality in isolation, knowing these endpoints aren't available in 'production'.

We want to add a `/check` endpoint for the changes we are about to make to the notice address generation, but we have identified some redundant ones that are no longer in use.

As a 'good team', we're deleting those first before adding our own!