Update notice setup journey (#2181)
* Update notice setup journey

https://eaflood.atlassian.net/browse/WATER-5143

We need to refactor the existing notice setup journey structure to accommodate the coming adHoc invitations and paper forms.

This change updates the previous 'journey' to be one of three types 'adHoc', 'standard' and 'abstraction-alerts'.

The previously known journeys of 'invitations' and 'reminders' are now notice types. This updates fits in with upcoming changes where the user will select which notice type from the select notice type page (currently selected via a link)