Add submit manual address service (#2180)
https://eaflood.atlassian.net/browse/WATER-5046

As part of the journey to create ad-hoc letters to be sent out to users we need to be able to select an address.

This PR adds in the functionality to submit the manual address page, validate it and save the information if validated, or display an error if not.