Update notification presenter to test adhoc templates
https://eaflood.atlassian.net/browse/WATER-5143

We recently changed the previous 'journey' key to be one of three types 'adHoc', 'standard' and 'abstraction-alerts'.

And the previously known journeys of 'invitations' and 'reminders' are now notice types. This updates fits in with upcoming changes where the user will select which notice type from the select notice type page (currently selected via a link)

This change adds the test to ensure an 'adhoc' 'invitation' uses the correct notify template.