Update session with data after insert
Currently, when we retrieve a session we update the returned object so that the content of `session.data` is elevated to the top level. However we don't do this when a session is inserted, so if we create a session and want to read from the returned session object, we must refer to the properties within `session.data` rather than the top level.

In a unit test for example, we may use `SessionHelper` to create a session containing a random condition id, which we later want to refer to in an assertion. We would have to refer to `session.data.conditionId` which is not consistent with the way we would refer to it in the actual code `session.conditionId`.

Elevating the properties within `session.data` is done by the method `SessionModel.$afterFind()`; we create an additional `SessionModel.$afterInsert()` which also does this, so that the session object returned after creating it will also have the properties elevated from `session.data` to the top level.