Refactor return version set up journey views (#2186)
https://eaflood.atlassian.net/browse/WATER-5139

Now that we have completed the return version set up journey it was time to revisit the views and standardise the code across them. As multiple people were working on the different pages it is always the case that there will be minor variations in how things are implemented.

This PR tries to standardise the approach across the views.