Amend `$afterFind()` unit tests
When we initially created our `$afterInsert()` tests, we cribbed the tets for `$afterFind()` since we're effectively testing for the same thing. We found the tests didn't quite work because we were checking the returned object keys by using `Object.keys(result)` to return an array and then comparing that array of keys with the expected array, and we found that the returned array was in a slightly different order.

We could have just changed the order of the array to fix this, but instead we changed the test to use the assertion `expect(result).to.only.include([...])`, which checks for the exact keys in any order. For consistency, we also make this change to the `$afterFind()` tests