Add code copied from the legacy water0abs-service
This is the code (tidied to get through our linter) that the legacy service uses to generate the address it sends to Notify, based on a contact from `crm.document_header.metadata.contacts`.

We're using it to generate what the legacy code does for comparison with what we are doing (and will do in the future).