Add a temp validate address service
Interestingly, we started by just attempting to preview a letter with each address. But we could not get Notify to fail the request, even though we were sure the address was invalid.

So, then we added in actually attempting to send a letter using the generated address. Now we started seeing in errors.