Add /check end points to verify Notify address fix (#2193)
https://eaflood.atlassian.net/browse/WATER-5147

We have a known issue with the addresses we are generating and sending to [GOV.UK Notify](https://www.notifications.service.gov.uk/).

If a contact has too many address fields populated, it results in us generating an address with more than seven lines, which Notify rejects.

We also want to verify that international addresses are handled correctly and explore various scenarios based on the address data we're finding.

So, we are making use of our non-production `/check` endpoints to add two new ones for testing.

- `/check/generate-address` will take a licence reference and return the original contact record, what Notify address the legacy code would have generated (for comparison) and what we are generating
- `/check/validate-address` will take a result from `/generate-address` and allow us to validate whether Notify will accept it or not

These are purely temporary, and only for testing and development of a fix. Hence the lack of test coverage and 'tidiness'!