Develop determine notify address for all scenarios
https://eaflood.atlassian.net/browse/WATER-5147

We have a known issue with the addresses we are generating and sending to [GOV.UK Notify](https://www.notifications.service.gov.uk/).

If a contact has too many address fields populated, it results in us generating an address with more than seven lines, which Notify rejects.

We also want to verify that international addresses are handled correctly and explore various scenarios based on the address data we're finding.

In [Add /check end points to verify Notify address fix](https://github.com/DEFRA/water-abstraction-system/issues/2193) we added some temporary endpoints to allow us to quickly see the addresses we're generating and validate whether Notify will accept them.

In that change we added `app/services/check/address/determine-notify-address.js` as representation of what we do now in the project.

This change develops a 'notify-address.presenter' that will be used to generate a Notify 'valid' address for a contact if possible, else it will highlight those that are invalid.