Create `formatValueUnit` base presenter function (#2195)
During QA of the tag a licence journey, it was noticed that we're not always consistent with how we display values with units; in some cases we display them with a space between the value and the unit (eg. `100 Ml/d`) and in other cases we display them without a space (eg. `100Ml/d`).

We've been advised that according to the GOV UK style guide, it's correct to not have a space between value and unit. In order to document this within our codebase, we create `formatValueUnit` in our base presenter which accepts a value and unit and displays them correctly, and we update our code wherever we find it to make use of this.