Add the new NotifyAddressPresenter
This was built by querying our address data for examples of the types of addresses we hold.

We then iterated the presenter to ensure it returned the address expected for each one.

In the next change we add the unit tests, which hopefully provide a more concrete example of the scenarios it needs to handle.