Allow zero when submitting a start reading (#2198)
https://eaflood.atlassian.net/browse/WATER-5167

An issue was found in production where a user wanted to put in a starting reading of 0 but we only allow positive numbers. This was due to assuming that all users already have a meter and thus the reading should always be positive however it is possible for new meters to be provided and they will start at zero.

This PR updates the validation for the start reading page to allow for 0.