Show zero value when revisiting page with saved session data (#2199)
https://eaflood.atlassian.net/browse/WATER-5167

An issue was found in production where a user wanted to put in a starting reading of 0 but we only allow positive numbers. This was due to assuming that all users already have a meter and thus the reading should always be positive however it is possible for new meters to be provided and they will start at zero.

This PR updates the presenter to show the 0 value when it has been saved by converting it to a string. When it is returned as a number it is seen as a falsely value and does not get displayed.