Fix missing address line in preview presenter
We can generate up to a maximum of 7 address lines, not 6 when we send a letter to Notify.

As part of the change we use the same filter-out nulls method used in other Notices services.

In the tests, we update the values to be more realistic. Name and address_line_1 will always match as they are set from the same source. If any address lines are null, they would be the last ones, not in the middle because even the previous way of generating the address filtered out null address parts.